requirements.txt
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.3
Django==2.2.2
djangorestframework==3.9.4
Dumper==1.2.0
factory-boy==2.12.0
Faker==1.0.7
httpie==1.0.2
idna==2.8
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
